+++ python/lynda-tutorial/section4syntax/syntax-conditionals.py
@@ -1,0 +1,13 @@
+#!/usr/bin/python3
+
+def main():
+    a, b = 0, 1
+    if a < b:
+        print("a is less than b")
+    elseif a > b:
+        print("a is greater than b")
+    else:
+        print("a is not less than b")
+
+if __name__ == "__main__": main()
+