+++ python/lynda-tutorial/section4syntax/syntax-functions.py
@@ -1,0 +1,11 @@
+#!/usr/bin/python3
+
+def main():
+    func()
+
+def func():
+    for i in range(10):
+        print(i, end=' ')
+    print()
+
+if __name__ == "__main__": main()