+++ python/lynda-tutorial/section4syntax/syntax-objects.py
@@ -1,0 +1,16 @@
+#!/usr/bin/python3
+
+class Egg:
+    def __init__(self, kind = "fried"):
+        self.kind = kind
+
+    def whatKind(self):
+        return self.kind
+
+def main():
+    fried = Egg()
+    scrambled = Egg('scrambled')
+    print(scrambled.whatKind())
+
+if __name__ == "__main__": main()
+