+++ python/lynda-tutorial/section4syntax/syntax-whitespace.py
@@ -1,0 +1,9 @@
+#!/usr/bin/python3
+
+def main():
+    print("This is the syntax.py file.")
+    print("This is another line")
+
+if __name__ == "__main__": main()
+
+