+++ python/lynda-tutorial/section4syntax/syntax-working.py
@@ -1,0 +1,9 @@
+#!/usr/bin/python3
+
+def main():
+    c = [1, 2, 3]
+    d = "Yo dawg! I heard you like Facebook?"
+    print(c)
+    print(d)
+
+if __name__ == "__main__": main()